Sun Studio warning, Converting an "int" into a 64-bit integer causes sign extension. Explicit cast resolves the issue.
git-svn-id: file:///home/asfernandes/fb/repository/sf/svn/firebird/trunk@45618 de594faa-8d1b-4a0c-9a6a-a7de5f8bf859

+++ src/jrd/perf.cpp
@@ -188,7 +188,7 @@ static int perf_format(const P* before, const P* after,
 	*p = 0;
 	const int length = p - buffer;
 	if (buffer_length && (buffer_length -= length) >= 0) {
-		memset(p, ' ', buffer_length);
+		memset(p, ' ', static_cast<size_t>(buffer_length));
 	}
 
 	return length;
